+++ R/standard_error.R
@@ -728,6 +728,19 @@ standard_error.complmrob <- function(model, ...) {
 
 
 #' @export
+standard_error.fixest <- function(model, ...) {
+  stats <- summary(model)
+  params <- insight::get_parameters(model)
+
+  .data_frame(
+    Parameter = params$Parameter,
+    SE = as.vector(stats$se)
+  )
+}
+
+
+
+#' @export
 standard_error.biglm <- function(model, ...) {
   cs <- summary(model)$mat
   params <- insight::get_parameters(model)